Plot phylogeny with predicted and known hosts

This function takes a phylogenetic tree (`phylo`) and a data frame containing information on predicted and known hosts for each tip, and produces a circular (or rectangular) tree with: - Predicted hosts highlighted with bold italic labels and custom colour. - Known hosts marked with a tip point and custom colour.

plot_phylo_hosts(
  phylo,
  pred_hosts,
  label,
  layout = "circular",
  tiplab_offset = 5,
  tiplab_size = 1.7,
  tippoint_shape = 18,
  tippoint_size = 2,
  pred_color = "darkgoldenrod3",
  nonpred_color = "gray50",
  known_color = "#cf3530",
  show_legend = FALSE
)

Arguments

phylo

An object of class `phylo` representing the tree to plot.

pred_hosts

A data.frame containing tip information with three columns: - `label`: character vector of tip names (species names matching tip labels). - `is_known`: logical vector indicating if the tip is a known host. - `is_pred`: logical vector indicating if the tip is a predicted host.

label

Character string indicating the column name in `pred_hosts` containing the tip labels.

layout

Character string specifying the tree layout (default: `circular`).

tiplab_offset

Numeric, offset distance of tip labels from the tree (defaut: `5`).

tiplab_size

Numeric, font size of tip labels (default: 1.7).

tippoint_shape

Numeric, shape code for known host tip points (default: `18`).

tippoint_size

Numeric, size of the tip point symbols (default: 2).

pred_color

Character, colour for predicted host labels (default: `darkgoldenrod3`).

nonpred_color

Character, colour for non-predicted tips (default: `gray50`).

known_color

Character, colour for known host tip points (default: `#cf3530`).

show_legend

Logical, whether to show a legend (default: `FALSE`).

Value

A `ggtree` object that can be printed or further customised.

Examples

pred_hosts <- data.frame(
  quercus_sp = oak_phylo$tip.label,
  is_known = oak_phylo$tip.label %in% known_hosts,
  is_pred = oak_phylo$tip.label %in% pred_hosts
)
#> Error: object 'oak_phylo' not found
plot_oak_phylo_hosts(oak_phylo, pred_hosts, label = "quercus_sp")
#> Error in plot_oak_phylo_hosts(oak_phylo, pred_hosts, label = "quercus_sp"): could not find function "plot_oak_phylo_hosts"