Plot a phylogeny highlighting host species for a focal species

This function plots a phylogenetic tree and highlights host species associated with a given focal species using tip label colours.

plot_phylogeny(
  phylo,
  host_spp = NULL,
  interaction_df = NULL,
  species = NULL,
  species_col = NULL,
  host_col = NULL,
  host_status_col = NULL,
  main = species,
  layout = "circular",
  offset = 5,
  size = 1.7,
  tip_colours = c("black", "coral"),
  legend_pos = "none"
)

Arguments

phylo

An object of class phylo containing the phylogenetic tree.

host_spp

A character string specifying the host species. This can be provided as an alternative to inferring the host species via `get_host_spp` when an `interaction_df` is supplied. Default: `NULL` (not required if `interaction_df` is provided).

interaction_df

A data frame containing species-host interaction data. Default: `NULL` (not required if `host_spp` is provided).

species

Character string giving the focal species of interest.

species_col

A character string indicating the column name in `interaction_df` that contains species names. Default: `NULL` (not required if `interaction_df` is provided).

host_col

A character string indicating the column name in both `model` and `interaction_df` that contains host species names.

host_status_col

A character string specifying the column in `interaction_df` that indicates the host status. Default: `NULL` (not required if `interaction_df` is provided).

main

Character string indicating plot title. Defaults to the name of the focal species (set to `NULL` to ignore).

layout

Character string specifying the tree layout passed to ggtree (e.g. "circular", "rectangular"). Default: `"circular"`.

offset

Number indicating the offset of tip labels from the phylogenetic tree. Default: `5`.

size

Number indicating the size of tip labels. Default: `1.7`.

tip_colours

Character vector of length 2, indicating colours for non-hosts and hosts. Default: `c("black", "coral")`.

legend_pos

Character string specifying legend position. Default: `"none"`.

Value

A ggplot object produced by ggtree.

See also

Examples

if (FALSE) { # \dontrun{
plot_phylogeny(phylo = oak_phylo,
               interaction_df = interaction_data,
               species = "Agrilus coxalis",
               species_col = "agrilus_sp",
               host_status_col = "interaction",
               host_col = "quercus_sp")
} # }